Altered Mapping to parameterCombinations
'extractDigitalDataML.m' has been modified to associate all good trial stimuli with the 4th dimension (spatial frequency) of the parameterCombinations cell, instead of the 5th (orientation). This brings it in line with the convention followed in previous projects working with images. Other minor code changes / improvements for compatibility with the Predictive Coding Project.